school-attendance/backend/requirements.txt
# Core Django
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
boto3==1.34.14

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4

# WebSocket & ASGI
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0

# Celery
celery==5.3.4
redis==5.0.1

# Face Recognition
face_recognition==1.3.0
dlib==19.24.2
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# Reports
reportlab==4.0.7
openpyxl==3.1.2

# Email
django-anymail==10.2

# Monitoring & Logging
django-prometheus==2.3.1

# Rate Limiting
django-ratelimit==4.1.0

# Utilities
python-decouple==3.8
requests==2.31.0
